Cleaning up App.js
- took out unnecessary functions from App.js and moved into proper folders
