+++ front-end/src/App.js
@@ -8,6 +8,7 @@ import {
   StorageTimeSearchListGrain,
   StorageTimeSearchListMeats,
 } from "./StorageTimeSearch/StorageTimeSearchListGroups";
+import StorageTimeSearchList from "./StorageTimeSearch/StorageTimeSearchList";
 import {StorageTimeSearch} from "./StorageTimeSearch/StorageTimeSearch";
 import {Recommendations} from "./Recommendations/Recommendations";
 import NavBar from "./NavBar";
@@ -15,8 +16,6 @@ import { Link, BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import MyFridge from "./MyFridge/MyFridge";
 import { ReadyToMake } from "./Recommendations/ReadyToMake";
 import { SavedRecipes } from "./Recommendations/SavedRecipes";
-import EditStorageModal from "./StorageTimeSearch/EditStorageTimeModal";
-import { useState } from 'react';
 
 // the app itself and the links for everything
 // the components are at the bottom and they are simply the pages and its contents
@@ -70,9 +69,6 @@ function App() {
   );
 }
 
-
-
-
 // the home page with the items and the stuff at the bottom
 const Home = () => (
   <div>
@@ -94,47 +90,5 @@ const Home = () => (
     </header>
   </div>
 );
-
-// the general structure for a storage time search list
-// includes a back link for every list
-export const StorageTimeSearchList = (props) => {
-  return (
-    <div>
-      <button className="StorageBackButton">
-        <Link to="/Storage-Time-Search">Back</Link>
-      </button>
-      <h1 className="Food-Group">{props.title}</h1>
-    </div>
-  );
-};
-
-// the general structure for every item within the list (button)
-// includes an item name, spoil time, and default time
-// uses a popup function when clicked
-export const StorageItem = ({ item, key }) => {
-  const [show, setShow] = useState(false)
-  return (
-    <div>
-      <button onClick={() => setShow(true)} className="Storage-Item">
-        {/* <img className="ItemStorageImage" src={item.img} alt='pic'></img> */}
-        <text className="Item-name">{item.name}</text>
-        <div className="All-time">
-          <text className="Spoil-time">Recommended</text>
-          <div className="Storage-Days">{item.shortTime}-{item.longTime} days</div>
-          <text className="Default-time">Default</text>
-          <div className="Storage-Days">{item.defaultTime} days</div>
-        </div>
-      </button>
-      <EditStorageModal title={item.name} onClose={() => setShow(false)} show={show}>
-        <ul>
-          <li>Shorter: {item.shortTime} days</li>
-          <li>Average: {item.averageTime} days</li>
-          <li>Longer: {item.longTime} days</li>
-        </ul>
-      </EditStorageModal>
-    </div>
-  )
-}
-
 
 export default App;
